Add pages and posts from avsitterplus.github.io

+++ clean.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-rm -rf _site .sass-cache .jekyll-cache
-rm -f .jekyll-metadata .DS_Store
+have_bundle=$(type -p bundle)
+if [ "${have_bundle}" ]; then
+  bundle exec jekyll clean
+else
+  rm -rf _site .sass-cache .jekyll-cache
+  rm -f .jekyll-metadata .DS_Store
+fi
